# More efficient parsing of json objects and arrays