Benchmark results using JDK 17 on Intel(R) Core(TM) i7-7700HQ CPU @ 2.80GHz (max 3.80GHz):

### Before
```
[info] Benchmark                                  (size)   Mode  Cnt         Score        Error  Units
[info] ADTReading.smithy4sJson                       N/A  thrpt    5    808749.173 ±  47973.445  ops/s
[info] AnyValsReading.smithy4sJson                   N/A  thrpt    5    722229.099 ±  52435.601  ops/s
[info] ArrayOfBigDecimalsReading.smithy4sJson        128  thrpt    5     78378.537 ±   7259.531  ops/s
[info] ArrayOfBigIntsReading.smithy4sJson            128  thrpt    5    100344.717 ±  15748.740  ops/s
[info] ArrayOfBooleansReading.smithy4sJson           128  thrpt    5    389046.699 ±  42920.955  ops/s
[info] ArrayOfBytesReading.smithy4sJson              128  thrpt    5    251991.326 ±  24871.638  ops/s
[info] ArrayOfDoublesReading.smithy4sJson            128  thrpt    5    146931.356 ±  21814.418  ops/s
[info] ArrayOfFloatsReading.smithy4sJson             128  thrpt    5    180607.384 ±   2788.820  ops/s
[info] ArrayOfInstantsReading.smithy4sJson           128  thrpt    5     73894.632 ±  10385.710  ops/s
[info] ArrayOfIntsReading.smithy4sJson               128  thrpt    5    225964.904 ±  30801.468  ops/s
[info] ArrayOfLongsReading.smithy4sJson              128  thrpt    5    209300.179 ±  11657.757  ops/s
[info] ArrayOfShortsReading.smithy4sJson             128  thrpt    5    259306.450 ±  10429.700  ops/s
[info] ArrayOfUUIDsReading.smithy4sJson              128  thrpt    5    207362.101 ±  25263.905  ops/s
[info] ArraySeqOfBooleansReading.smithy4sJson        128  thrpt    5    372376.820 ±  45837.375  ops/s
[info] ExtractFieldsReading.smithy4sJson             128  thrpt    5     99799.603 ±   3776.470  ops/s
[info] GeoJSONReading.smithy4sJson                   N/A  thrpt    5     12297.252 ±   1506.092  ops/s
[info] GitHubActionsAPIReading.smithy4sJson          N/A  thrpt    5    147309.098 ±   1133.314  ops/s
[info] GoogleMapsAPIReading.smithy4sJson             N/A  thrpt    5      5919.545 ±    717.013  ops/s
[info] ListOfBooleansReading.smithy4sJson            128  thrpt    5    813130.930 ± 106372.686  ops/s
[info] MapOfIntsToBooleansReading.smithy4sJson       128  thrpt    5     75870.305 ±   2193.631  ops/s
[info] MissingRequiredFieldsReading.smithy4sJson     N/A  thrpt    5   1895044.882 ± 175041.064  ops/s
[info] NestedStructsReading.smithy4sJson             128  thrpt    5     40489.538 ±   5012.136  ops/s
[info] OpenRTBReading.smithy4sJson                   N/A  thrpt    5     61004.648 ±   8780.130  ops/s
[info] PrimitivesReading.smithy4sJson                N/A  thrpt    5    728054.743 ± 106550.844  ops/s
[info] SetOfIntsReading.smithy4sJson                 128  thrpt    5     92656.363 ±   7107.091  ops/s
[info] TwitterAPIReading.smithy4sJson                N/A  thrpt    5     14395.426 ±    748.829  ops/s
[info] VectorOfBooleansReading.smithy4sJson          128  thrpt    5    872798.975 ±  19769.464  ops/s
```
### After
```
[info] Benchmark                                  (size)   Mode  Cnt         Score        Error  Units
[info] ADTReading.smithy4sJson                       N/A  thrpt    5   1092372.241 ± 227090.250  ops/s
[info] AnyValsReading.smithy4sJson                   N/A  thrpt    5   1252623.855 ± 241810.578  ops/s
[info] ArrayOfBigDecimalsReading.smithy4sJson        128  thrpt    5     87774.292 ±   8240.544  ops/s
[info] ArrayOfBigIntsReading.smithy4sJson            128  thrpt    5    112687.106 ±   9683.479  ops/s
[info] ArrayOfBooleansReading.smithy4sJson           128  thrpt    5    519264.014 ±  77951.406  ops/s
[info] ArrayOfBytesReading.smithy4sJson              128  thrpt    5    347772.443 ±  39887.564  ops/s
[info] ArrayOfDoublesReading.smithy4sJson            128  thrpt    5    162470.127 ±  19084.472  ops/s
[info] ArrayOfFloatsReading.smithy4sJson             128  thrpt    5    211907.875 ±  24799.776  ops/s
[info] ArrayOfInstantsReading.smithy4sJson           128  thrpt    5     82877.423 ±    655.298  ops/s
[info] ArrayOfIntsReading.smithy4sJson               128  thrpt    5    286152.975 ±  44965.615  ops/s
[info] ArrayOfLongsReading.smithy4sJson              128  thrpt    5    248874.665 ±  47774.826  ops/s
[info] ArrayOfShortsReading.smithy4sJson             128  thrpt    5    331409.539 ±  54126.152  ops/s
[info] ArrayOfUUIDsReading.smithy4sJson              128  thrpt    5    238240.147 ±  46342.773  ops/s
[info] ArraySeqOfBooleansReading.smithy4sJson        128  thrpt    5    483484.457 ±  61963.462  ops/s
[info] ExtractFieldsReading.smithy4sJson             128  thrpt    5    114567.326 ±    291.743  ops/s
[info] GeoJSONReading.smithy4sJson                   N/A  thrpt    5     19161.752 ±   2110.739  ops/s
[info] GitHubActionsAPIReading.smithy4sJson          N/A  thrpt    5    228375.333 ±  17807.735  ops/s
[info] GoogleMapsAPIReading.smithy4sJson             N/A  thrpt    5      9055.412 ±    358.508  ops/s
[info] ListOfBooleansReading.smithy4sJson            128  thrpt    5    968133.598 ±  38316.824  ops/s
[info] MapOfIntsToBooleansReading.smithy4sJson       128  thrpt    5    104793.528 ±  14893.378  ops/s
[info] MissingRequiredFieldsReading.smithy4sJson     N/A  thrpt    5   3321600.850 ± 211941.759  ops/s
[info] NestedStructsReading.smithy4sJson             128  thrpt    5     66419.963 ±    154.514  ops/s
[info] OpenRTBReading.smithy4sJson                   N/A  thrpt    5     98103.451 ±    658.625  ops/s
[info] PrimitivesReading.smithy4sJson                N/A  thrpt    5   1242346.976 ± 183669.824  ops/s
[info] SetOfIntsReading.smithy4sJson                 128  thrpt    5    138854.811 ±   2675.555  ops/s
[info] TwitterAPIReading.smithy4sJson                N/A  thrpt    5     20042.600 ±    180.671  ops/s
[info] VectorOfBooleansReading.smithy4sJson          128  thrpt    5   1028548.455 ±   5580.704  ops/s
```